# batch/tassidicambio/target/generated-sources/openapi/src/main/java/com/example/restservice/api/CurrenciesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.restservice.api;

import com.example.restservice.model.Currency;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-05-09T14:26:53.298332800+02:00[Europe/Rome]")
@Validated
@Controller
@Tag(name = "Exchange Rates", description = "the Exchange Rates API")
public interface CurrenciesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /currencies : Visualizza la lista delle divise ISO.
     * 
     *
     * @return successful operation (status code 200)
     *         or Default error sample response (status code 200)
     */
    @Operation(
        operationId = "viewCurrencies",
        summary = "Visualizza la lista delle divise ISO.",
        description = "",
        tags = { "Exchange Rates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Currency.class))),
                @Content(mediaType = "application/xml", array = @ArraySchema(schema = @Schema(implementation = Currency.class)))
            }),
            @ApiResponse(responseCode = "default", description = "Default error sample response")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/currencies",
        produces = { "application/json", "application/xml" }
    )
    
    default ResponseEntity<List<Currency>> viewCurrencies(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"country\" : \"STATI UNITI\", \"currencyISO\" : \"USD\", \"validityStartDate\" : \"1918-01-02T00:00:00.000+00:00\", \"currencyName\" : \"Dollaro USA\", \"countryISO\" : \"US\", \"tmsInsert\" : \"2000-01-23T04:56:07.000+00:00\", \"validityEndDate\" : \"2000-01-23\" }, { \"country\" : \"STATI UNITI\", \"currencyISO\" : \"USD\", \"validityStartDate\" : \"1918-01-02T00:00:00.000+00:00\", \"currencyName\" : \"Dollaro USA\", \"countryISO\" : \"US\", \"tmsInsert\" : \"2000-01-23T04:56:07.000+00:00\", \"validityEndDate\" : \"2000-01-23\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/xml"))) {
                    String exampleString = "<currency> <currencyISO>USD</currencyISO> <currencyName>Dollaro USA</currencyName> <country>STATI UNITI</country> <countryISO>US</countryISO> <validityStartDate>Wed Jan 02 01:00:00 CET 1918</validityStartDate> <validityEndDate>2000-01-23</validityEndDate> <tmsInsert>2000-01-23T04:56:07.000Z</tmsInsert> </currency>";
                    ApiUtil.setExampleResponse(request, "application/xml", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}