# batch/tassidicambio/target/generated-sources/openapi/src/main/java/com/example/restservice/api/DailyExchangeRatesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.restservice.api;

import com.example.restservice.model.ExchangeRate;
import com.example.restservice.model.ExchangeRatesRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-05-09T14:26:53.298332800+02:00[Europe/Rome]")
@Validated
@Controller
@Tag(name = "Exchange Rates", description = "the Exchange Rates API")
public interface DailyExchangeRatesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /dailyExchangeRates : Visualizza, rispetto alla valuta e alla data di riferimento inserite in request, i tassi di cambio di tutte le divise passate in input.
     * 
     *
     * @param exchangeRatesRequest  (optional)
     * @return successful operation (status code 200)
     *         or Default error sample response (status code 200)
     */
    @Operation(
        operationId = "viewDailyExchangeRates",
        summary = "Visualizza, rispetto alla valuta e alla data di riferimento inserite in request, i tassi di cambio di tutte le divise passate in input.",
        description = "",
        tags = { "Exchange Rates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ExchangeRate.class))),
                @Content(mediaType = "application/xml", array = @ArraySchema(schema = @Schema(implementation = ExchangeRate.class)))
            }),
            @ApiResponse(responseCode = "default", description = "Default error sample response")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/dailyExchangeRates",
        produces = { "application/json", "application/xml" },
        consumes = { "application/json", "application/xml", "application/x-www-form-urlencoded" }
    )
    
    default ResponseEntity<List<ExchangeRate>> viewDailyExchangeRates(
        @Parameter(name = "ExchangeRatesRequest", description = "") @Valid @RequestBody(required = false) ExchangeRatesRequest exchangeRatesRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"country\" : \"AUSTRALIA\", \"exchangeConvention\" : \"Quantità di valuta estera per 1 Euro\", \"referenceCurrency\" : \"EUR\", \"isoCode\" : \"AUD\", \"uicCode\" : 109, \"tmsInsert\" : \"2000-01-23T04:56:07.000+00:00\", \"avgRate\" : 1.6515, \"currency\" : \"Dollaro Australiano\", \"referenceDate\" : \"2024-02-26T00:00:00.000+00:00\" }, { \"country\" : \"AUSTRALIA\", \"exchangeConvention\" : \"Quantità di valuta estera per 1 Euro\", \"referenceCurrency\" : \"EUR\", \"isoCode\" : \"AUD\", \"uicCode\" : 109, \"tmsInsert\" : \"2000-01-23T04:56:07.000+00:00\", \"avgRate\" : 1.6515, \"currency\" : \"Dollaro Australiano\", \"referenceDate\" : \"2024-02-26T00:00:00.000+00:00\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/xml"))) {
                    String exampleString = "<exchangeRate> <country>AUSTRALIA</country> <currency>Dollaro Australiano</currency> <isoCode>AUD</isoCode> <uicCode>109</uicCode> <avgRate>1.6515</avgRate> <exchangeConvention>Quantità di valuta estera per 1 Euro</exchangeConvention> <referenceDate>Mon Feb 26 01:00:00 CET 2024</referenceDate> <referenceCurrency>EUR</referenceCurrency> <tmsInsert>2000-01-23T04:56:07.000Z</tmsInsert> </exchangeRate>";
                    ApiUtil.setExampleResponse(request, "application/xml", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}